move CIFAR loading code from example to main library
